refactor: move ReceiptComponent to a new location and update imports accordingly
