+++ app/src/components/sections/ExpenseAccountView/__tests__/TransactionHistorySection.spec.ts
@@ -3,7 +3,7 @@ import { flushPromises, mount } from '@vue/test-utils'
 import TransactionHistorySection from '@/components/sections/ExpenseAccountView/TransactionHistorySection.vue'
 import type { ExpenseTransaction } from '@/types/transactions'
 import ButtonUI from '@/components/ButtonUI.vue'
-import ReceiptComponent from '@/components/sections/ExpenseAccountView/ReceiptComponent.vue'
+import ReceiptComponent from '@/components/ReceiptComponent.vue'
 import TableComponent from '@/components/TableComponent.vue'
 import { ref } from 'vue'
 import { createTestingPinia } from '@pinia/testing'
